remember original replace directive
When we run `gohack get` on a go.mod file that contains a replace
directive, we now remember the old information inside a comment,
and restore it when running `gohack undo`.

To do this, we need to be able to parse go.mod comments, so we
now use the modfile package rather than `go mod edit`.

We also use testscript to add some tests. More tests still needed :)